refactor(cli): split chat share add command files
